requirements.txt
contentful==1.3.0
Flask==0.10.1
Frozen-Flask==0.12
jinja2==2.8
markdown==2.6.5
pygments==2.0.2
pypugjs==4.2.2
pyyaml==3.11
requests==2.7

# blog
ipython[notebook]
tortilla
